+++ src/editor/components/menu/view-mode-toggle-button.jsx
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import { connect } from "react-redux";
 import Button from "@material-ui/core/Button";
 import Tooltip from "@material-ui/core/Tooltip";
-import { setViewMode } from "../../actions/actions";
+import { setViewMode } from "../../actions/notebook-actions";
 
 export class ViewModeToggleButtonUnconnected extends React.Component {
   static propTypes = {