+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   reuse in multiple server environments (#1943)
 - Hide "unsaved changes" in revision browser if no unsaved changes (#2046)
 - Show revision, docs, contribute links even on non-public sites
+- Disable "History" menu item for the tryit notebook
+- Add documentation for how the Iodide server's authentication system works
 
 # 0.10.0 (2019-07-15)